Convert cli calls to rest (#17)
* Convert to fasthttp from cli calls

* Update README with further instructions

* Update the test use cases to use the internal REST calls instead of cli packages

* Update error message to give more information

* Fix missing utils reference

* Fix logic for calls and tests

* Change wrap to wrapf where sprintf is used for error messages
Add in jpillora/backoff logic for relay pulls and project deletes to offset the case where the cluster creation occurs faster than the relays can populate when calling getbootstrapfile back. And the case where the cluster resource still exists even after a delete command is requested, making it impossible for the project to be deleted. This is to avoid orphan projects.

* Update registry docs

* Fix explaination of attribute

---------

Co-authored-by: Ethan Stein <ethan.stein@iherb.com>